updating delete to validate password and fleshing out error handling

+++ src/api/accounts/use-cases/delete/delete-account.handler.ts
@@ -1,6 +1,8 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { AccountRepository } from '../../repository/account.repository';
 import { DeleteAccountCommand } from './delete-account.command';
+import * as bcrypt from 'bcrypt';
+import { Result } from '@badrap/result';
 
 @CommandHandler(DeleteAccountCommand)
 export class DeleteAccountHandler
@@ -8,6 +10,21 @@ export class DeleteAccountHandler
 {
   constructor(private readonly accountRepository: AccountRepository) {}
   async execute(command: DeleteAccountCommand): Promise<any> {
+    const desiredAccountToUpdate = await this.accountRepository.findById(
+      command.id,
+    );
+
+    if (desiredAccountToUpdate.isErr) {
+      return Result.err(new NoRecordAvailable());
+    }
+    const isValidPassword = await bcrypt.compare(
+      command.password,
+      desiredAccountToUpdate.unwrap().password,
+    );
+
+    if (!isValidPassword) {
+      return Result.err(new InvalidPassword());
+    }
     const accountDeleteResult = await this.accountRepository.delete(command.id);
     return accountDeleteResult.map(
       (deletedAccount) => deletedAccount,
@@ -19,3 +36,7 @@ export class DeleteAccountHandler
 export abstract class AccountDeleteErrorResponse extends Error {}
 
 export class RepositoryDeleteError extends AccountDeleteErrorResponse {}
+
+export class InvalidPassword extends AccountDeleteErrorResponse {}
+
+export class NoRecordAvailable extends AccountDeleteErrorResponse {}
